chore(mypy): promote web/ + crypto cycle/scheduler/components to --strict
Item 5 follow-up — pulls the dashboard route layer and the three
load-bearing crypto composition files into the gated mypy
``files`` list. ``uv run mypy`` now checks 174 source files clean
(up from 103) and CI gates them on the same strict settings as
``domain/`` and ``core/``.

Key seam changes the promotion required:
* ``LLMBackend`` Protocol (domain/ports.py) gains
  ``generate_tool_call`` — every concrete LLM (BaseLLM subclasses,
  FallbackLLM) already exposes it, so the Protocol now matches
  runtime. Wave H's agentic loop and the strategy's tool-use call
  site (``_call_tool``) can therefore type against the Protocol
  rather than the concrete BaseLLM class.
* ``BaseCycleService._bus`` typed as ``EventBus | None`` so the
  subclass annotations and the ``run_stages(bus=…)`` callers
  type-check. ``_publish_event`` reads ``self._bus`` directly
  instead of going through ``getattr`` (the cast was needed only
  because ``_bus`` was ``object | None``).
* ``CryptoTradingStrategy.ensemble_llms`` typed as
  ``Sequence[LLMBackend]`` (covariant) instead of
  ``list[LLMBackend]`` (invariant) so the composition root can
  hand in a ``list[BaseLLM]`` without an explicit cast.
* ``timed_broker_call`` decorator (core/metrics.py) given
  ``ParamSpec`` + ``TypeVar`` typing so wrapped broker methods
  still satisfy the ``Broker`` / ``CancellableBroker`` Protocols
  after the histogram wrapping (previously the wrapper's
  ``(*Any, **Any) -> Awaitable[Any]`` shape broke Protocol
  compatibility for ``cancel_order`` and ``get_open_orders``).

Web-layer cleanups (real bugs, not just paperwork):
* web/metrics.py: ``by_provider`` bucket was typed as
  ``dict[str, float | int]`` but stored a list under
  ``elapsed_ms_list`` — fixed to ``dict[str, Any]`` and tightened
  the token-coercion path so a non-numeric ``tokens`` field is
  dropped instead of polluting the int counter.
* web/partnership_directory.py: removed a dead-code branch the
  static analyser flagged (``to_stage is PAUSED`` was already
  caught by the outer ``if``).
* web/routes/insights.py: validate ``status`` against the
  ``Literal['pending', 'approved', 'rejected', 'deferred']``
  before calling ``ExceptionQueue.decide`` instead of relying on
  a ``# type: ignore[arg-type]``.
* web/routes/admin_halal.py: guard ``avg_entry_price`` against
  ``None`` so the ``float()`` coercion has a valid input.

Tests: 11536 pass, 18 pre-existing failures unchanged. The 99
stale-``app_state.clear()`` errors (Wave A residue) are now
fixed across 11 test files — the corresponding fixture lines
were already no-ops since Wave A deleted the global state shim.

+++ src/halal_trader/web/accessibility_audit.py
@@ -101,7 +101,7 @@ def _channel_luminance(channel_byte: int) -> float:
     c = channel_byte / 255.0
     if c <= 0.03928:
         return c / 12.92
-    return ((c + 0.055) / 1.055) ** 2.4
+    return float(((c + 0.055) / 1.055) ** 2.4)
 
 
 def relative_luminance(color: str) -> float:
